requirements.txt
Flask
python-dotenv
ibm-watson-machine-learning
